requirements.txt
streamlit==1.41.0
psycopg2-binary==2.9.9
streamlit-lottie
requests
